MovieItem to common moved
