+++ src/common/MovieItem/styled.js
